requirements.txt
gradio
easyocr
opencv-python-headless
Pillow
numpy
